Refactor: Introduce update screen, reverse geocoding, and data model improvements
This commit introduces a major refactoring and several new features:

-   **Update Screen:**
    -   Added a new `UpdateScreen` to edit a person's details (first name, last name, address).
    -   Created `UpdateViewModel` and `UpdateEvent` to manage the state and user interactions on the update screen.
    -   Navigation is updated to route from the list and map info windows to the new `UpdateScreen`, passing the person's ID.

-   **Data Model & Database:**
    -   Modified the `Person` and `PersonEntity` data models:
        -   Changed `id` from `Long` to `Int`.
        -   Made `lat` and `lng` nullable to support persons without a location.
        -   Added an `address` field for storing reverse-geocoded location strings.
    -   Updated the Room database schema to reflect these changes.

-   **Map & Geocoding:**
    -   Implemented reverse geocoding in `MapsViewModel` to automatically fetch and store a formatted address when a new location is added via a long-press on the map.
    -   When a new person is created from the map, it now navigates directly to the `UpdateScreen` for that new person.
    -   Markers on the map now display the person's address in the snippet.

-   **List View:**
    -   The list page now displays the person's address.
    -   Added a dialog to inform the user when they try to view a person on the map who does not have a location set.
    -   Created a dedicated `ListViewModel`.

-   **Configuration:**
    -   Set up API key handling using `apikey.properties` (added to `.gitignore`) and manifest placeholders to securely manage Google Maps API keys for debug and release builds.

+++ app/src/main/java/com/lucwaw/friendsLocal/ui/Detail.kt
@@ -5,6 +5,6 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 
 @Composable
-fun DetailPage(personId: Long?, modifier: Modifier = Modifier) {
+fun DetailPage(personId: Int?, modifier: Modifier = Modifier) {
     Text(text = "Detail Page - PersonId = ${personId ?: "None"}", modifier = modifier)
 }
